+++ libmcompiler/mcompiler/compiler.h
@@ -8,13 +8,15 @@
 #ifndef _COMPILER_H
 #define _COMPILER_H
 
+#include <stdint.h>
+
 int compiler_verlib(char *ver);
 
 typedef unsigned short sym_t;
 
 struct lex_s	{
 	sym_t		sym;
-	unsigned long	data;
+	uintptr_t	data;
 		};
 
 #define	COMPILE_TAB_TERMINALS	7
